EconomySnapshot struct header file added.
