+++ Source/ClickerGame/Private/Data/Economy/EconomySnapshot.cpp
@@ -1,0 +1,2 @@
+#include "Data/Economy/EconomySnapshot.h"
+
