+++ docs/diary/Devlog.md
@@ -479,3 +479,30 @@
   - 2-3) 중복 선언으로 인한 중복 구독으로 발생하는 에러임을 인지하고 수정하여 해결
 - proof
   - Commit Link
+
+## 09.25.25.
+- what
+  1. FloatingText 풀 한도 초과시 설정해둔 경고 메시지 발생 확인 
+  2. EconomySnapshot 구조체 독립
+  3. README 링크 일부 수정
+  4. Delegate 선언 관련 솔루션 오류 해결
+  5. 프로젝트 내 함수 및 멤버 포인터 플로우 파악
+- why
+  - 1-1) 풀을 초과할 정도의 토스트를 생성할 때 대비책이 필요하다고 생각.
+  - 2-1) 유지보수 측면과 가독성 측면에서 EconomySubsystem.h에 기생하는 구조면 합리적이지 못하다고 판단.
+  - 3-1) README에 작성한 하위파일 링크가 파일명 대소문자 구분으로 정상 작동하지 않음. 포트폴리오로 올린 목적에 반하는 상황이기에 수정 결심.
+  - 4-1) EconomySnapshot 파일 추가 도중 솔루션 빌드에 오류발생
+  - 4-2) 솔루션 빌드 실패 원인은 델리게이트 인자 이름 중복
+  - 5-1) 프로젝트 전반을 파악하며 포인터 플로우 이해의 부족을 체감하여 더 파악하기로 결정.
+- how
+  - 1-1) 문제점 인지한 상태로 마감.
+  - 2-1) EconomySnapshot.h를 Public/Data/Economy/ 경로에 생성, 콜드빌드로 솔루션에 캐싱 후 내용 작성
+  - 2-2) 에디터 실행 후 데이터 정상작동 확인.
+  - 3-1) Github에서 파일명 직접 수정하여 커밋
+  - 4-1) 빌드 실패 로그로 델리게이트 관련 오류 내용 파악
+  - 4-2) 델리게이트 매개변수 이름 중복부분 확인
+  - 4-3) 수정 후 재 빌드 성공
+  - 5-1) GPT와 포인터 관련 블로그 게시글을 활용하여 포인터 플로우 이해, 파악
+- proof
+  - Commit Link
+  - 1. Image ![Image](./assets/FloatingText_warning.png)
