+++ docs/planning/Daily/2025-09/2025-09-25.md
@@ -1,0 +1,32 @@
+# 2025-09-25
+
+## Topic
+
+- Economy Idle/Floating 위젯 OnFinished 반납 리팩터
+
+## Tasks
+
+- [x] FEconomySnapshot 선언 스크립트 분리
+- [ ] 1Tick, Offline 재화지급 루틴 델리게이트로 수정
+- [ ] Floating/Idle Text 생성/반납 리팩터
+- [ ] Toast Widget C++클래스로 공통기능 상속
+
+## Blockers
+
+- None
+
+## Notes
+
+- FloatingText 위젯 풀 한도 초과 시 작성해둔 경고문구 출력 확인
+  - 한도 초과 시 대응책 마련 필요
+
+## EOD
+
+- Done:
+  - FEconomySnapshot 선언 스크립트 분리
+- Evidence:
+  - Commit Link
+- Carry to Tomorrow:
+  - 1Tick, Offline 재화지급 루틴 델리게이트로 수정
+  - Floating/Idle Text 생성/반납 루틴 리팩터
+  - Toast Widget C++ 클래스로 공통 기능 상속
