+++ docs/planning/Daily/2025-09/2025-09-26.md
@@ -1,0 +1,28 @@
+# 2025-09-26
+
+## Topic
+
+- Economy Idle/Floating 위젯 OnFinished 반납 리팩터
+
+## Tasks
+
+- [ ] 1Tick, Offline 재화지급 루틴 델리게이트로 수정
+- [ ] Floating/Idle Text 생성/반납 리팩터
+- [ ] Toast Widget C++클래스로 공통기능 상속
+
+## Blockers
+
+- None
+
+## Notes
+
+- 
+
+## EOD
+
+- Done:
+  - 
+- Evidence:
+  - Commit Link
+- Carry to Tomorrow:
+  -
